@@ -20,7 +20,7 @@ import Foundation
 ///     }
 /// }
 /// ```
-public protocol LoggerHandler {
+public protocol LoggerHandler: Sendable {
     /// Handles a log event.
     ///
     /// - Parameters:
@@ -35,12 +35,12 @@ public protocol LoggerHandler {
     func log(
         level: LogLevel,
         message: String,
-        parameters: [String: Any],
+        parameters: [String: Sendable],
         category: LogCategory,
         module: String,
         file: String,
         line: Int,
         function: String
-    )
+    ) async
 }
 
