@@ -5,13 +5,16 @@
 //  Created by Evgeniy Yolchev on 05.11.2025.
 //
 
-import XCTest
+import Testing
+import Foundation
 @testable import Logger
 @testable import LoggerImpl
 
-final class LogMessageFormatterTests: XCTestCase {
+@Suite
+struct LogMessageFormatterTests {
     
-    func testFormatLogMessageWithAllArguments() {
+    @Test
+    func formatLogMessageWithAllArguments() {
         let message = formatLogMessage(
             level: .warning,
             message: "Test message",
@@ -25,13 +28,14 @@ final class LogMessageFormatterTests: XCTestCase {
         
         let lines = message.split(separator: "\n")
         
-        XCTAssertEqual(lines.count, 3)
-        XCTAssertTrue(lines[0].hasSuffix(" [WARNING] [TestModule] [Network] Test message"))
-        XCTAssertEqual(String(lines[1]), "Parameters: key=value")
-        XCTAssertEqual(String(lines[2]), "Location: TestFile.swift:123 testFunc()")
+        #expect(lines.count == 3)
+        #expect(lines[0].hasSuffix(" [WARNING] [TestModule] [Network] Test message"))
+        #expect(String(lines[1]) == "Parameters: key=value")
+        #expect(String(lines[2]) == "Location: TestFile.swift:123 testFunc()")
     }
     
-    func testFormatLogMessageWithEmptyParameters() {
+    @Test
+    func formatLogMessageWithEmptyParameters() {
         let message = formatLogMessage(
             level: .info,
             message: "Simple message",
@@ -45,13 +49,14 @@ final class LogMessageFormatterTests: XCTestCase {
         
         let lines = message.split(separator: "\n")
         
-        XCTAssertEqual(lines.count, 3)
-        XCTAssertTrue(lines[0].hasSuffix(" [INFO] [Module] [Domain] Simple message"))
-        XCTAssertEqual(String(lines[1]), "Parameters: none")
-        XCTAssertEqual(String(lines[2]), "Location: File.swift:5 func()")
+        #expect(lines.count == 3)
+        #expect(lines[0].hasSuffix(" [INFO] [Module] [Domain] Simple message"))
+        #expect(String(lines[1]) == "Parameters: none")
+        #expect(String(lines[2]) == "Location: File.swift:5 func()")
     }
     
-    func testFormatLogMessageWithMultipleParameters() {
+    @Test
+    func formatLogMessageWithMultipleParameters() {
         let message = formatLogMessage(
             level: .debug,
             message: "Request completed",
@@ -65,18 +70,19 @@ final class LogMessageFormatterTests: XCTestCase {
         
         let lines = message.split(separator: "\n")
         
-        XCTAssertEqual(lines.count, 3)
-        XCTAssertTrue(lines[0].hasSuffix(" [DEBUG] [APIClient] [Network] Request completed"))
+        #expect(lines.count == 3)
+        #expect(lines[0].hasSuffix(" [DEBUG] [APIClient] [Network] Request completed"))
         
         let parametersLine = String(lines[1])
-        XCTAssertTrue(parametersLine.hasPrefix("Parameters: "))
-        XCTAssertTrue(parametersLine.contains("status=200"))
-        XCTAssertTrue(parametersLine.contains("duration=1.5"))
+        #expect(parametersLine.hasPrefix("Parameters: "))
+        #expect(parametersLine.contains("status=200"))
+        #expect(parametersLine.contains("duration=1.5"))
         
-        XCTAssertEqual(String(lines[2]), "Location: API.swift:25 handleResponse()")
+        #expect(String(lines[2]) == "Location: API.swift:25 handleResponse()")
     }
     
-    func testFormatLogMessageTimestampFormat() {
+    @Test
+    func formatLogMessageTimestampFormat() {
         let message = formatLogMessage(
             level: .critical,
             message: "Test",
@@ -96,7 +102,7 @@ final class LogMessageFormatterTests: XCTestCase {
         let range = NSRange(message.startIndex..., in: message)
         let match = timestampRegex.firstMatch(in: message, options: [], range: range)
         
-        XCTAssertNotNil(match)
+        #expect(match != nil)
     }
 }
 
