@@ -4,13 +4,22 @@ final class FileLoggerHandler: LoggerHandler {
     func log(
         level: LogLevel,
         message: String,
-        parameters: [String : Any],
+        parameters: [String: Sendable],
         category: LogCategory,
         module: String,
         file: String,
         line: Int,
-        function: String
+        function: String,
     ) {
-        
+        let formattedMessage = formatLogMessage(
+            level: level,
+            message: message,
+            parameters: parameters,
+            category: category,
+            module: module,
+            file: file,
+            line: line,
+            function: function
+        )
     }
 }
