@@ -45,7 +45,7 @@ final class OSLoggerHandler: LoggerHandler {
     func log(
         level: LogLevel,
         message: String,
-        parameters: [String: Any],
+        parameters: [String: Sendable],
         category: LogCategory,
         module: String,
         file: String,
